linux: enable support for dri/drm
With this we enable support for dri/drm AND keep support for the simple
framebuffer interface. (CONFIG_DRM_SIMPLEDRM instead of
CONFIG_FB_SIMPLE)

This increases the complexity and size of the kernel and introduces some
overhead. However, some newer applications need drm, and the run-time
overhead should be negligible.