buildroot: linux: renamed linux simpleinput to simpleinputptr (moved to mouse)
Simpleinput was put under misc input devices before, because we
originally planned planned to support mouse AND keyboard with it.
Now we plan to use dedicated drivers for mouse and keyboard, so we
renamed simpleinput to simpleinputptr and put it under mouse drivers.
In near future a separate new driver simpleinputkbd will be added to
support keyboard input.