Efficient boot: switch rootfs from initramfs to romfs using physmap drivers
Using the newly introduced memory mapped file in the vp, we can now
switch from rootfs as initramfs to rootfs images provided by block
devices. This leads to much shorter system start-up times and to less
main memory usage. The image is restricted to a size of 512MB for RV64
and, due to limited vmalloc size, to 64MB for RV32.

The change includes:
 * Buildroot:
   * rootfs: Create romfs images instead of cpio/initramfs
   * Added the mtdinfo tool (for debugging)
 * Kernel:
   * Disable support for initramfs
   * Added support for MTD subsystem, MTD Block devices and physmap
     drivers (mtd-ram) -> Memory mapped file to linux block device
 * Devicetree
   * Added mtd-ram node corresponding to mapping in the VP
   * Changed boot parameters regarding rootfs mount
 * Build/Run System
   * Prevent modification of buildroot results: rootfs.romfs is copied
     to runtime_mram/mram_root.img before use by the vp
   * Added vp parameters to map runtime_mram/mram_root.img

Using this mechanism can lead to much faster system start-up times (~1/2
time to login compared to initramfs) and more free main memory.