Persistent Storage: add read-write datafs(ext4) using physmap drivers
Similar as it was done for rootfs, the vp now supports a second memory mapped
file for persistent storage (data).

With this, we realize a very common pattern in embedded linux system design:
A read-only root filesystem to hold system and constant data, and a read-write
data file for persistent / generated data.

As with the rootfs, the image is restricted to a size of 512MB for RV64
and, due to limited vmalloc size, to 64MB for RV32.

The change includes:
 * Buildroot:
   * Support for the ext4 tools (filesystem creation and check)
   * Bootscript to initialize, check and mount /data
 * Kernel:
   * Support for the ext4 filesystem.
 * Devicetree
   * Added mtd-ram node corresponding to mapping of data in the VP
 * Build/Run System
   * Added vp parameters to map runtime_mram/mram_rv32_data.img and
     runtime_mram/mram_rv64_data.img

After starting the vp, the ext4 filesystem images runtime_mram/mram_rv32_data.img
are available (mounted) und /data