pre-tag 2024.12 (fix RISC-V VP++ to TAG 2024.12)
There are pending changes in RISC-V VP++ that will break compatibility
with GUI-VP Kit. In order to be able to create permanently reproducible
setups (GUI-VP Kit + RISC-V VP++), the RISC-V VP++ version to be used
is explicitly fixed to TAG "2024.12" in this change. In addition, a TAG of
the same name will be created for GUI-VP Kit.