buildroot: bump to OpenSBI version 1.6 (CAUTION: COMPATIBILITY)
CAUTION: For RV32, this change does only work with RISC-V VP++ versions
>= git hash de8d3d0189! (more on that below)

Newer OpenSBI versions are relocatable. As a side-effect, the load address
(0x80000000) is no longer hard-coded in OpenSBI and therefore also not set
in the resulting ELF file.
In short: The VP loads OpenSBI at the wrong address.

To fix this, we explicitly specify the load-address with the optional
OpenSBI build parameter FW_TEXT_START to 0x80000000.
With this OpenSBI is loaded to the correct address on RV32 and RV64
again.
However, the boot of OpenSBI still failed on RV32 due to a bug in
RISC-V VP++ (missing mstatush csr on RV32).
This bug is fixed in RISC-V VP++ with commit
de8d3d018969e6f2d22778f6263a0eb30616caa8
"vp: core: rv32: add missing csr mstatush and missing bits (ube, sbe, mbe)"
Please make sure to use at least this version of RISC-V VP++.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>