dt: switch to device tree provided by RISC-V VP++
With this we switch to the device-tree generator "linux-dt-gen.py"
introduced in RISC-V VP++ in
https://github.com/ics-jku/riscv-vp-plusplus.git
with
858a63389ce81d89af94560e5e400e76e5a44303
"vp: platform: linux: added the device tree (dts) generator "linux-dt-gen.py""

"linux-dt-gen.py" is based on "dt/gen_dts.sh", which is removed by this
change. However, the new version is written in python and also allows to
set the memory start address and bootargs.

Moving the device tree to RISC-V VP++ improves overall compatibility of
GUI-VP Kit with different versions of RISC-V VP++.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>